flask-docgen/requirements.txt
flask
gitpython
openai

